Refactor: HistroyList에서 eventHandler 분리
