+++ src/layout/Header/Section/SearchBox/History/index.js
@@ -1,15 +1,9 @@
 import HtmlElement from '../../../../../utils/HtmlElement.js';
-import { moveToSearchTermPage } from '../../../../../router.js';
-import { myLocalStorage } from '../../../../../utils/mockDB.js';
 import { SEARCH_BOX } from '../../../../../constant.js';
+import eventHandler from './eventHandler.js';
 
 const {
-  HISTORY: {
-    HISTORY_DELETE,
-    HISTORY_ACTIVE,
-    HISTORY_DELETE__ALL,
-    HISTORY_LOCAL_STORAGE_KEY,
-  },
+  HISTORY: { HISTORY_DELETE, HISTORY_ACTIVE, HISTORY_DELETE__ALL },
 } = SEARCH_BOX;
 
 export default function HistoryList($element, args) {
@@ -18,6 +12,13 @@ export default function HistoryList($element, args) {
 
 HistoryList.prototype = Object.create(HtmlElement.prototype);
 HistoryList.prototype.constructor = HistoryList;
+
+HtmlElement.prototype.init = function () {
+  this.state = {
+    ...this.args,
+  };
+  this.eventHandler = eventHandler;
+};
 
 HistoryList.prototype.setTemplate = function () {
   const { histroyList, activeHistory } = this.state;
@@ -46,38 +47,6 @@ HistoryList.prototype.setTemplate = function () {
 };
 
 HistoryList.prototype.setEvent = function () {
+  const { handleClick } = this.eventHandler;
   this.$element.addEventListener('click', handleClick.bind(this));
 };
-
-function handleClick({ target }) {
-  const {
-    dataset: { clickType },
-  } = target;
-  switch (clickType) {
-    case HISTORY_DELETE:
-      deleteTargetTerm.call(this, target);
-      break;
-    case HISTORY_ACTIVE:
-      moveToSearchTermPage(this.state.option, target.innerText.slice(0, -1));
-      break;
-    case HISTORY_DELETE__ALL:
-      deleteAllTerm.apply(this);
-    default:
-      break;
-  }
-}
-
-function deleteAllTerm() {
-  myLocalStorage.set(HISTORY_LOCAL_STORAGE_KEY, []);
-  this.setState({ histroyList: [] });
-}
-
-function deleteTargetTerm(target) {
-  const updatedHistroyList = [...this.state.histroyList];
-  const {
-    dataset: { termId: targetTermId },
-  } = target.closest('li');
-  updatedHistroyList.splice(targetTermId, 1);
-  myLocalStorage.set(HISTORY_LOCAL_STORAGE_KEY, updatedHistroyList);
-  this.setState({ histroyList: updatedHistroyList });
-}
